add support for different filtered views in topology
