+++ deepfence_frontend/apps/dashboard/src/api/api.ts
@@ -47,9 +47,11 @@ export function getUserApiClient() {
 export function getTopologyApiClient() {
   const topologyApi = new TopologyApi(configuration);
   return {
+    getCloudTopologyGraph: topologyApi.getTopologyGraph.bind(topologyApi),
     getHostsTopologyGraph: topologyApi.getHostsTopologyGraph.bind(topologyApi),
     getKubernetesTopologyGraph: topologyApi.getKubernetesTopologyGraph.bind(topologyApi),
-    getCloudTopologyGraph: topologyApi.getTopologyGraph.bind(topologyApi),
+    getContainersTopologyGraph: topologyApi.getContainersTopologyGraph.bind(topologyApi),
+    getPodsTopologyGraph: topologyApi.getPodsTopologyGraph.bind(topologyApi),
   };
 }
