+++ deepfence_frontend/apps/dashboard/src/features/topology/types/graph.ts
@@ -101,3 +101,5 @@ export type TopologyAction =
   | {
       type: 'refresh';
     };
+
+export type TopologyViewType = 'cloud' | 'kubernetes' | 'host' | 'pod' | 'container';
